fix: update permissions and queries
