+++ api/src/apollo/EventRegistration/mutations.ts
@@ -38,6 +38,7 @@ export const deleteEventRegistration = mutationField(
         PERMISSIONS.SUPER_ADMIN,
         PERMISSIONS.SUPER_EDITOR,
         PERMISSIONS.ORG_ADMIN,
+        PERMISSIONS.ORG_EDITOR,
       ]),
     args: {
       id: nonNull(idArg()),
