+++ api/src/apollo/Scalars/GenderType.ts
@@ -1,0 +1,7 @@
+import { enumType } from 'nexus';
+
+export const GenderType = enumType({
+  name: 'GenderType',
+  members: ['MALE', 'FEMALE', 'OTHERS'],
+  asNexusMethod: 'gender',
+});
