+++ api/src/apollo/Scalars/StatusType.ts
@@ -1,0 +1,7 @@
+import { enumType } from 'nexus';
+
+export const StatusType = enumType({
+  name: 'StatusType',
+  members: ['ACTIVE', 'DRAFT', 'EXPIRED'],
+  asNexusMethod: 'status',
+});
