+++ api/src/apollo/Scalars/TransactionType.ts
@@ -1,0 +1,7 @@
+import { enumType } from 'nexus';
+
+export const TransactionType = enumType({
+  name: 'TransactionType',
+  members: ['REGISTRATION', 'MERCH', 'EVENT'],
+  asNexusMethod: 'transactionType',
+});
