+++ api/src/apollo/index.ts
@@ -1,4 +1,5 @@
-export * from './scalers';
+// export * from './scalers';
+export * from './Scalars';
 export * from './User';
 export * from './Org';
 export * from './Event';
